+++ rust/main/agents/relayer/src/relayer.rs
@@ -41,7 +41,7 @@ use crate::{
         pending_message::{MessageContext, MessageSubmissionMetrics},
         processor::{MessageProcessor, MessageProcessorMetrics},
     },
-    server::{self as relayer_server, MessageRetryRequest},
+    server::{self as relayer_server},
     settings::{matching_list::MatchingList, RelayerSettings},
 };
 use crate::{
@@ -311,7 +311,7 @@ impl BaseAgent for Relayer {
                 }));
             tasks.push(console_server.instrument(info_span!("Tokio console server")));
         }
-        let sender = BroadcastSender::<MessageRetryRequest>::new(ENDPOINT_MESSAGES_QUEUE_SIZE);
+        let sender = BroadcastSender::<MatchingList>::new(ENDPOINT_MESSAGES_QUEUE_SIZE);
         // send channels by destination chain
         let mut send_channels = HashMap::with_capacity(self.destination_chains.len());
         let mut prep_queues = HashMap::with_capacity(self.destination_chains.len());