+++ rust/main/chains/hyperlane-sealevel/src/lib.rs
@@ -15,6 +15,7 @@ pub use solana_sdk::signer::keypair::Keypair;
 pub use trait_builder::*;
 pub use validator_announce::*;
 
+mod account;
 mod error;
 mod interchain_gas;
 mod interchain_security_module;
@@ -24,4 +25,6 @@ mod multisig_ism;
 mod provider;
 mod rpc;
 mod trait_builder;
+mod transaction;
+mod utils;
 mod validator_announce;